refactor(frontend): Update import statements
